requirements.txt

streamlit>=1.36.0
python-dotenv>=1.0.1
openai>=1.35.3
requests>=2.32.3
faiss-cpu>=1.8.0
pandas>=2.2.2
numpy>=1.26.4
scikit-learn>=1.4.2
